(SIMP-1542) Support for SIMP 6
This updates the commands to work on either SIMP 6 or the Legacy SIMP
systems by pulling all information from puppet itself and the
surrounding operating system.

When the older versions of SIMP are retired, this should be modified to
use the AIO Ruby instead of the system Ruby.

SIMP-1542 #close
SIMP-1540 #comment Updated 'simp-cli' for SIMP 6 with backward compatibility